Added multiple users to exist on one instance. Updated some of the css. Bumped version of crypto for security reasons (#10)

+++ persistence/stickyConnection.go
@@ -6,32 +6,48 @@ import (
 
 // Dat sticky sticky
 type StickyConnection struct {
-	RouteID          string    `json:"id,omitempty"`
-	ConnectionApiKey string    `json:"connectionKey,omitempty"`
-	TopicName        string    `json:"topicName,omitempty"`
-	LastModified     time.Time `json:"lastModified,omitempty"`
+	RouteID      string    `json:"id,omitempty"`
+	Security_key string    `json:"securityKey,omitempty"`
+	TopicName    string    `json:"topicName,omitempty"`
+	LastModified time.Time `json:"lastModified,omitempty"`
+	UserId       int       `json:"userId,omitempty"`
 }
 
-func (dstore *Datastore) GetStickeyConnection(id string) (StickyConnection, error) {
-	selectQuery := "SELECT id, connection_api_key, topic_name, last_modified FROM stickey_connections WHERE id=?;"
+func (dstore *Datastore) GetStickeyConnectionBySecurityKey(id string, security_key string) (StickyConnection, error) {
+	selectQuery := "SELECT id, security_key, topic_name, last_modified, user_id FROM stickey_connections WHERE id=? and security_key=?;"
 	dstore.RWMutex.RLock()
-	row := dstore.db.QueryRow(selectQuery, id)
+	row := dstore.db.QueryRow(selectQuery, id, security_key)
 	dstore.RWMutex.RUnlock()
 
 	var stickey_connection StickyConnection
 
-	err := row.Scan(&stickey_connection.RouteID, &stickey_connection.ConnectionApiKey, &stickey_connection.TopicName, &stickey_connection.LastModified)
+	err := row.Scan(&stickey_connection.RouteID, &stickey_connection.Security_key, &stickey_connection.TopicName, &stickey_connection.LastModified)
 	if err != nil {
 		return stickey_connection, err
 	}
 	return stickey_connection, nil
 }
 
-func (dstore *Datastore) GetAllStickyConnections() ([]StickyConnection, error) {
-	stickey_connections := []StickyConnection{}
-	selectQuery := "SELECT id, connection_api_key, topic_name, last_modified FROM stickey_connections;"
+func (dstore *Datastore) GetStickeyConnectionByUser(id string, userId int) (StickyConnection, error) {
+	selectQuery := "SELECT id, security_key, topic_name, last_modified, user_id FROM stickey_connections WHERE id=? and user_id=?;"
 	dstore.RWMutex.RLock()
-	rows, err := dstore.db.Query(selectQuery)
+	row := dstore.db.QueryRow(selectQuery, id, userId)
+	dstore.RWMutex.RUnlock()
+
+	var stickey_connection StickyConnection
+
+	err := row.Scan(&stickey_connection.RouteID, &stickey_connection.Security_key, &stickey_connection.TopicName, &stickey_connection.LastModified, &stickey_connection.UserId)
+	if err != nil {
+		return stickey_connection, err
+	}
+	return stickey_connection, nil
+}
+
+func (dstore *Datastore) GetAllStickyConnections(userId int) ([]StickyConnection, error) {
+	stickey_connections := []StickyConnection{}
+	selectQuery := "SELECT id, security_key, topic_name, last_modified FROM stickey_connections WHERE user_id=?;"
+	dstore.RWMutex.RLock()
+	rows, err := dstore.db.Query(selectQuery, userId)
 	dstore.RWMutex.RUnlock()
 	if err != nil {
 		return stickey_connections, err
@@ -40,20 +56,19 @@ func (dstore *Datastore) GetAllStickyConnections() ([]StickyConnection, error) {
 
 	for rows.Next() {
 		var stickey_connection StickyConnection
-		err := rows.Scan(&stickey_connection.RouteID, &stickey_connection.ConnectionApiKey, &stickey_connection.TopicName, &stickey_connection.LastModified)
+		err := rows.Scan(&stickey_connection.RouteID, &stickey_connection.Security_key, &stickey_connection.TopicName, &stickey_connection.LastModified)
 		if err != nil {
 			return stickey_connections, err
 		}
 		stickey_connections = append(stickey_connections, stickey_connection)
 	}
-
 	return stickey_connections, nil
 }
 
 func (dstore *Datastore) InsertStickyConnection(sticky_connection StickyConnection) (int, error) {
-	insertQuery := "INSERT INTO stickey_connections(id, connection_api_key, topic_name) VALUES(?, ?, ?);"
+	insertQuery := "INSERT INTO stickey_connections(id, security_key, topic_name, user_id) VALUES(?, ?, ?, ?);"
 	dstore.RWMutex.Lock()
-	res, err := dstore.db.Exec(insertQuery, sticky_connection.RouteID, sticky_connection.ConnectionApiKey, sticky_connection.TopicName)
+	res, err := dstore.db.Exec(insertQuery, sticky_connection.RouteID, sticky_connection.Security_key, sticky_connection.TopicName, sticky_connection.UserId)
 	dstore.RWMutex.Unlock()
 	if err != nil {
 		return 0, err
@@ -65,10 +80,10 @@ func (dstore *Datastore) InsertStickyConnection(sticky_connection StickyConnecti
 	return int(rowsCreated), nil
 }
 
-func (dstore *Datastore) DeleteStickyConnection(stickey_connection_id string) (int, error) {
-	insertQuery := "DELETE FROM stickey_connections WHERE id=?"
+func (dstore *Datastore) DeleteStickyConnection(stickey_connection_id string, userId int) (int, error) {
+	insertQuery := "DELETE FROM stickey_connections WHERE id=? AND user_id=?"
 	dstore.RWMutex.Lock()
-	res, err := dstore.db.Exec(insertQuery, stickey_connection_id)
+	res, err := dstore.db.Exec(insertQuery, stickey_connection_id, userId)
 	dstore.RWMutex.Unlock()
 	if err != nil {
 		return 0, err
